agent/requirements.txt
# Backend
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
geoalchemy2>=0.14.0

# AI & LLM
langgraph>=0.0.60
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
geojson==3.1.0
shapely==2.0.2

# Frontend
streamlit==1.28.2
plotly==5.17.0
folium==0.15.0
streamlit-folium==0.15.0
pydeck==0.8.1

# HTTP & Async
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: Vector Store
faiss-cpu==1.7.4
qdrant-client==1.7.0

# Optional: Additional ML
scikit-learn==1.3.2
transformers==4.36.0
